requirements.txt
# Core AI dependencies
transformers==4.49.0
bitsandbytes==0.45.3
qwen-vl-utils==0.0.10
accelerate>=0.26.0

# Data processing
pandas==2.2.3
numpy==1.26.3
scikit-learn==1.5.2
scipy==1.14.1
regex==2024.11.6
rapidfuzz==3.10.1

# Web framework
fastapi==0.115.4
starlette==0.41.2
uvicorn==0.32.0

# Image processing
Pillow==11.1.0
opencv-python-headless==4.10.0.84

# Utility and support
jinja2==3.1.3
python-dotenv==1.0.1
filelock==3.17.0
click==8.1.7
setuptools-scm==8.2.0
termcolor==2.5.0

# JSON and parsing
jsonschema==4.23.0
protobuf==3.20.2

# Multipart form data support
python-multipart==0.0.17